Change 1) AudioBus is now embeded in the AudioClipInfoClass class. The default value is set OnEnable (hope this is the right way to do... ideally everytime they need to be this way, I don't think any specific character will have a different priority for a specific "trigger"
Because of this change, the "AudioBus priority" parameter  in the AudioManagerMk2.PlaySound should now be obsolete (see image)
I've renamed the parameter to "priorityObsolete" but would suggest to remove from all instances of the method. Since it's a big change, I've not proceeded with this one myself. In case we agree on all the changes, I can do that.

Change 2) Added the actual Audio buses (not only the priority system but the actual mixer with the side chaining. Now the mix should be a little clearer, High prio is lowering volume on mid_prio and low_prio and mid is lowering volume on low_prio.
Music is never affected.
Please let me know how it feels

Change 3) Added the fade in time in the AudioManagerMk2.PlaySound method.

I've used this specifically for the charge in of the Power Charge but have defaulted to 0.0f so no other modification is needed.